requirements.txt
boto3>=1.35.0
botocore>=1.35.0
certifi>=2024.7.4
charset-normalizer>=3.3.0
cryptography>=42.0.0
fastapi>=0.110.0
groq>=0.9.0
httpx>=0.27.0
keyring>=25.0.0
langchain>=0.1.0
langchain-classic>=1.0.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.30
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langchain-groq>=0.1.0
langchain-ollama>=0.1.0
langchain-openai>=0.1.0
langchain-perplexity>=1.0.0
langgraph>=0.1.0
mcp>=0.1.0
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
uvicorn>=0.27.0
